# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hiring Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Position {
  id: number;
  title: string;
  department: string;
  location: string;
  /** full-time, part-time, contract, internship */
  type: string;
  /** open, paused, closed */
  status: string;
  /** @nullable */
  description?: string | null;
  headcount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePositionBody {
  title: string;
  department: string;
  location: string;
  type: string;
  status: string;
  /** @nullable */
  description?: string | null;
  headcount: number;
}

export interface UpdatePositionBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  headcount?: number | null;
}

export interface Candidate {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  positionId?: number | null;
  /** @nullable */
  positionTitle?: string | null;
  /** @nullable */
  department?: string | null;
  /** applied, screening, technical, hr_interview, final_interview, offer, hired, rejected, withdrawn */
  stage: string;
  /** active, on_hold, archived */
  status: string;
  /**
   * linkedin, referral, job_board, direct, agency
   * @nullable
   */
  source?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /**
   * 1-5 overall rating
   * @nullable
   */
  rating?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCandidateBody {
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  positionId?: number | null;
  stage: string;
  status: string;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rating?: number | null;
}

export interface UpdateCandidateBody {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  positionId?: number | null;
  /** @nullable */
  stage?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rating?: number | null;
}

export interface Interview {
  id: number;
  candidateId: number;
  /** @nullable */
  candidateName?: string | null;
  /** @nullable */
  positionTitle?: string | null;
  /** screening, technical, hr_interview, final_interview, case_study */
  type: string;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  interviewerName?: string | null;
  /** @nullable */
  interviewerEmail?: string | null;
  /**
   * video_call, phone, on_site
   * @nullable
   */
  location?: string | null;
  /** @nullable */
  meetingUrl?: string | null;
  /** scheduled, completed, cancelled, no_show */
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateInterviewBody {
  candidateId: number;
  type: string;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  interviewerName?: string | null;
  /** @nullable */
  interviewerEmail?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  meetingUrl?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateInterviewBody {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  interviewerName?: string | null;
  /** @nullable */
  interviewerEmail?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  meetingUrl?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Feedback {
  id: number;
  interviewId: number;
  candidateId: number;
  /** @nullable */
  candidateName?: string | null;
  /** @nullable */
  interviewType?: string | null;
  reviewerName: string;
  /** @nullable */
  reviewerEmail?: string | null;
  /** 1-5 overall rating */
  rating: number;
  /** @nullable */
  technicalScore?: number | null;
  /** @nullable */
  culturalScore?: number | null;
  /** @nullable */
  communicationScore?: number | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** strong_yes, yes, neutral, no, strong_no */
  recommendation: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateFeedbackBody {
  interviewId: number;
  candidateId: number;
  reviewerName: string;
  /** @nullable */
  reviewerEmail?: string | null;
  rating: number;
  /** @nullable */
  technicalScore?: number | null;
  /** @nullable */
  culturalScore?: number | null;
  /** @nullable */
  communicationScore?: number | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  recommendation: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateFeedbackBody {
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  technicalScore?: number | null;
  /** @nullable */
  culturalScore?: number | null;
  /** @nullable */
  communicationScore?: number | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** @nullable */
  recommendation?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardSummary {
  totalCandidates: number;
  activeCandidates: number;
  hiredThisMonth: number;
  openPositions: number;
  interviewsScheduled: number;
  pendingFeedback: number;
  offersPending: number;
  /**
   * Average days from application to hire
   * @nullable
   */
  avgTimeToHire?: number | null;
}

export interface PipelineStat {
  stage: string;
  count: number;
  label: string;
}

export interface ActivityItem {
  id: number;
  /** candidate_added, stage_changed, interview_scheduled, feedback_submitted, offer_sent, hired, rejected */
  type: string;
  description: string;
  /** @nullable */
  candidateName?: string | null;
  /** @nullable */
  positionTitle?: string | null;
  timestamp: string;
}

export type ListCandidatesParams = {
  /**
   * @nullable
   */
  positionId?: number | null;
  /**
   * @nullable
   */
  stage?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
};

export type ListInterviewsParams = {
  /**
   * @nullable
   */
  candidateId?: number | null;
};

export type ListFeedbackParams = {
  /**
   * @nullable
   */
  interviewId?: number | null;
  /**
   * @nullable
   */
  candidateId?: number | null;
};