# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hiring Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - name: positionId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: stage
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCandidateBody"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCandidateBody"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /feedback:
    get:
      operationId: listFeedback
      tags: [feedback]
      summary: List all feedback
      parameters:
        - name: interviewId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: candidateId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"
    post:
      operationId: createFeedback
      tags: [feedback]
      summary: Submit feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackBody"
      responses:
        "201":
          description: Created feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
  /feedback/{id}:
    patch:
      operationId: updateFeedback
      tags: [feedback]
      summary: Update feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedbackBody"
      responses:
        "200":
          description: Updated feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
    delete:
      operationId: deleteFeedback
      tags: [feedback]
      summary: Delete feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        positionId:
          type: ["integer", "null"]
        positionTitle:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        stage:
          type: string
          description: "applied, screening, technical, hr_interview, final_interview, offer, hired, rejected, withdrawn"
        status:
          type: string
          description: "active, on_hold, archived"
        source:
          type: ["string", "null"]
          description: "linkedin, referral, job_board, direct, agency"
        resumeUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
          description: "1-5 overall rating"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - stage
        - status
        - createdAt
        - updatedAt
    CreateCandidateBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        positionId:
          type: ["integer", "null"]
        stage:
          type: string
        status:
          type: string
        source:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
      required:
        - firstName
        - lastName
        - email
        - stage
        - status
    UpdateCandidateBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        positionId:
          type: ["integer", "null"]
        stage:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
...
    Feedback:
      type: object
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        interviewType:
          type: ["string", "null"]
        reviewerName:
          type: string
        reviewerEmail:
          type: ["string", "null"]
        rating:
          type: integer
          description: "1-5 overall rating"
        technicalScore:
          type: ["integer", "null"]
        culturalScore:
          type: ["integer", "null"]
        communicationScore:
          type: ["integer", "null"]
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        recommendation:
          type: string
          description: "strong_yes, yes, neutral, no, strong_no"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - interviewId
        - candidateId
        - reviewerName
        - rating
        - recommendation
        - createdAt
    CreateFeedbackBody:
      type: object
      properties:
        interviewId:
          type: integer
        candidateId:
          type: integer
        reviewerName:
          type: string
        reviewerEmail:
          type: ["string", "null"]
        rating:
          type: integer
        technicalScore:
          type: ["integer", "null"]
        culturalScore:
          type: ["integer", "null"]
        communicationScore:
          type: ["integer", "null"]
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        recommendation:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - interviewId
        - candidateId
        - reviewerName
        - rating
        - recommendation
    UpdateFeedbackBody:
      type: object
      properties:
        rating:
          type: ["integer", "null"]
        technicalScore:
          type: ["integer", "null"]
        culturalScore:
          type: ["integer", "null"]
        communicationScore:
          type: ["integer", "null"]
        strengths:
          type: ["string", "null"]
        weaknesses:
          type: ["string", "null"]
        recommendation:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...