# lib/api-zod/src/generated/types/createCandidateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hiring Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCandidateBody {
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  positionId?: number | null;
  stage: string;
  status: string;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  rating?: number | null;
}