# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hiring Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: number;
  interviewId: number;
  candidateId: number;
  /** @nullable */
  candidateName?: string | null;
  /** @nullable */
  interviewType?: string | null;
  reviewerName: string;
  /** @nullable */
  reviewerEmail?: string | null;
  /** 1-5 overall rating */
  rating: number;
  /** @nullable */
  technicalScore?: number | null;
  /** @nullable */
  culturalScore?: number | null;
  /** @nullable */
  communicationScore?: number | null;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  /** strong_yes, yes, neutral, no, strong_no */
  recommendation: string;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}